src/main/java/com/epam/reportportal/jobs/service/project/ProjectService.java
package com.epam.reportportal.jobs.service.project;

import com.epam.reportportal.jobs.repository.ProjectRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * @author <a href="mailto:ivan_budayeu@epam.com">Ivan Budayeu</a>
 */
@Service
public class ProjectService {

	private final ProjectRepository projectRepository;

	@Autowired
	public ProjectService(ProjectRepository projectRepository) {
		this.projectRepository = projectRepository;
	}

	public List<Long> getAllIds() {
		return projectRepository.selectIds();
	}
}
